sort of finished... fixed up data collection. markov chains work mostly, but still some bugs. weird sound issues. nonetheless, it doesnt sound great but it was a good attempt

+++ src/com/nullpandaexception/locality/CameraData.java
@@ -4,9 +4,9 @@ import org.opencv.core.Mat;
 
 public class CameraData {
     public Mat freqs;
-    public double[] colors;
+    public float[] colors;
     
-    public CameraData(Mat f, double[] c) {
+    public CameraData(Mat f, float[] c) {
         freqs = f;
         colors = c;
     }
