+++ src/com/nullpandaexception/locality/GeoData.java
@@ -1,10 +1,10 @@
 package com.nullpandaexception.locality;
 
 public class GeoData {
-    public double longitude;
-    public double latitude;
+    public float longitude;
+    public float latitude;
     
-    public GeoData(double lon, double lat) {
+    public GeoData(float lon, float lat) {
         longitude = lon;
         latitude = lat;
     }
